+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Misc:
 - **HAML-Lint** Enable `parallel` option by default [#2012](https://github.com/sider/runners/pull/2012)
 - Verify gem installation in Dockerfiles [#2010](https://github.com/sider/runners/pull/2010)
 - **Slim-Lint** New support [#2014](https://github.com/sider/runners/pull/2014)
+- **RuboCop** Set up default config only when no user config [#2020](https://github.com/sider/runners/pull/2020)
 
 ## 0.41.1